X11: Split off the event loop into a separate module
