+++ src/x11/mod.rs
@@ -5,5 +5,6 @@ mod window;
 pub use window::*;
 
 mod cursor;
+mod event_loop;
 mod keyboard;
 mod visual_info;
